LayoutTests/http/tests/security/contentSecurityPolicy/extensions/manifest-v3/worker-src/manifest-v3-worker-src-block-partial-wildcard-expected.txt

CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/contentSecurityPolicy/resources/worker.py?type=post-message-pass&csp=script-src%20%27self%27%20%27unsafe-inline%27 because it does not appear in the worker-src directive of the Content Security Policy.
ALERT: SecurityError: The operation is insecure.
Test that CSP for manifest V3 extensions blocks partial wildcards for the worker-src directive. Test passes if we alert with an error.
